Move conditions test to ConditionLayout
